# docs/OPENAPI.yaml
...
openapi: 3.1.0
info:
  title: Oblivion API
  description: 'Backend API for controlled data erasure, recovery, residual analysis, assurance and certificate
    verification.


    This document is GENERATED by scripts/gen_openapi.py from the FastAPI application itself. Do not hand-edit it:
    edit the routes and schemas, then regenerate.


    Authentication uses HTTP Bearer with opaque server-side session tokens (Argon2id password hashing, SHA-256 token
    digest stored server-side). These are NOT JWTs. There is no refresh-token endpoint and no cookie based session;
    the bearer token is the only credential transport.'
  version: 0.1.0
paths:
...
  /api/certificates/{certificate_id}:
    get:
      tags:
      - certificates
      summary: Get Certificate
      description: 'Retrieve an issued certificate (requires evidence.view).


        Returns the signer''s public key only; no private key exists in this API.'
      operationId: get_certificate_api_certificates__certificate_id__get
      parameters:
      - name: certificate_id
        in: path
        required: true
        schema:
          type: string
          title: Certificate Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/certificates/{certificate_id}/verify:
    post:
      tags:
      - certificates
      summary: Verify Certificate Endpoint
      description: 'Verify a certificate across ten independent dimensions.


        The caller may supply the operation and target it *expects* this certificate

        to attest to. It cannot supply the outcome: signer trust comes from the

        server''s TrustStore, and evidence integrity is recomputed from the stored

        evidence. A request that supplies no expectations gets ``INCONCLUSIVE``,

        which is the truthful answer when nothing independent was available to check

        against.'
      operationId: verify_certificate_endpoint_api_certificates__certificate_id__verify_post
      parameters:
      - name: certificate_id
        in: path
        required: true
        schema:
          type: string
          title: Certificate Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/CertificateVerificationRequest'
              - type: 'null'
              title: Request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateVerificationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CertificateOut:
      properties:
        id:
          type: string
          title: Id
        operation_id:
          type: string
          title: Operation Id
        evidence_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Id
        evidence_digest:
          type: string
          title: Evidence Digest
        certificate_version:
          type: string
          title: Certificate Version
        canonicalization_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Canonicalization Version
        evidence_schema_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Schema Version
        target_identity:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Identity
        method:
          anyOf:
          - type: string
          - type: 'null'
          title: Method
        result:
          anyOf:
          - type: string
          - type: 'null'
          title: Result
        signer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Signer Id
        signing_algorithm:
          type: string
          title: Signing Algorithm
        key_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Key Id
        public_key:
          type: string
          title: Public Key
        signature:
          type: string
          title: Signature
        claim:
          anyOf:
          - type: string
          - type: 'null'
          title: Claim
        issued_at:
          type: string
          format: date-time
          title: Issued At
      type: object
      required:
      - id
      - operation_id
      - evidence_digest
      - certificate_version
      - signing_algorithm
      - public_key
      - signature
      - issued_at
      title: CertificateOut
    CertificateVerificationOut:
      properties:
        certificate_id:
          type: string
          title: Certificate Id
        overall_status:
          type: string
          pattern: ^(VALID|INVALID|INCONCLUSIVE)$
          title: Overall Status
          enum:
          - VALID
          - INVALID
          - INCONCLUSIVE
        dimensions:
          items:
            $ref: '#/components/schemas/DimensionResultOut'
          type: array
          title: Dimensions
          minItems: 10
          description: 'Exactly one result per verification dimension. A PASS on SIGNATURE_VALIDITY does not imply
            overall certificate validity: only overall_status does. NOT_CHECKED and INCONCLUSIVE are honest absences
            of evidence and must not be rendered as success.'
        cannot_prove:
          items:
            type: string
          type: array
          title: Cannot Prove
        certificate_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Certificate Version
        verifier_version:
          type: string
          title: Verifier Version
        verified_at:
          type: string
          format: date-time
          title: Verified At
        signature_valid:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Signature Valid
        evidence_integrity:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Evidence Integrity
        signer_trusted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Signer Trusted
      type: object
      required:
      - certificate_id
      - overall_status
      - dimensions
      - verifier_version
      - verified_at
      title: CertificateVerificationOut
    CertificateVerificationRequest:
      properties:
        expected_operation_id:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Expected Operation Id
          description: The operation the caller believes this certificate attests to. Compared against the certificate
            and its evidence.
        expected_target_identity:
          anyOf:
          - type: string
            maxLength: 1024
          - type: 'null'
          title: Expected Target Identity
          description: The target the caller believes this certificate attests to. Compared against the certificate
            and its evidence.
      additionalProperties: false
      type: object
      title: CertificateVerificationRequest
      description: 'Independent expectations the caller brings to the verification.


        Both are optional. Omitting one is honest - the corresponding dimension

        reports ``NOT_CHECKED`` and the overall result becomes ``INCONCLUSIVE``

        rather than pretending the check succeeded.'
...
    DimensionResultOut:
      properties:
        dimension:
          type: string
          title: Dimension
          enum:
          - STRUCTURE
          - VERSION_COMPATIBILITY
          - EVIDENCE_AVAILABILITY
          - EVIDENCE_DIGEST
          - SIGNATURE_VALIDITY
          - PUBLIC_KEY_CONSISTENCY
          - SIGNER_TRUST
          - EVIDENCE_CHAIN_INTEGRITY
          - OPERATION_CONSISTENCY
          - TARGET_CONSISTENCY
        result:
          type: string
          pattern: ^(PASS|FAIL|NOT_CHECKED|INCONCLUSIVE)$
          title: Result
          enum:
          - PASS
          - FAIL
          - NOT_CHECKED
          - INCONCLUSIVE
        detail:
          type: string
          title: Detail
        evidence_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Ref
      type: object
      required:
      - dimension
      - result
      - detail
      title: DimensionResultOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpaqueSessionToken
      description: 'Opaque session token issued by POST /api/auth/login. Send as ''Authorization: Bearer <token>''.
        The server stores only the SHA-256 digest and revokes it on logout.'
security:
- bearerAuth: []